azure-functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
playwright==1.38.0
beautifulsoup4==4.9.3
requests==2.26.0
azure-storage-blob==12.23.0